cross out items, show seperately items

+++ src/App.js
@@ -3,11 +3,12 @@ import React, { useState, useEffect, useRef} from 'react'
 export default function App() {
     var form = useRef()
     var item_l = useRef()
-    var done = useRef()
-    var deleteItem = useRef()
 
     const [todoList, setTodoList] = useState([]);
-    const [wholeList, setWholeList] = useState([]);
+    const [crossOutList, setCrossOutList] = useState([]);
+    const [showDone, setShowDone] = useState(false);
+    const [showAll, setShowAll] = useState(true);
+    const [showNotDone, setShowNotDone] = useState(false)
 
     function add_todo() {
         var nItem = form.current.value
@@ -32,6 +33,35 @@ export default function App() {
         }))
     }
 
+    function cross_item(num) {
+        var value = todoList[num];
+        if(crossOutList.includes(value)) {
+            setCrossOutList(prevList => prevList.filter((item) => {
+                return item != value;
+            }))
+        } else {
+            setCrossOutList(prevList => [...prevList, value])
+        }
+    }
+
+    function show_done() {
+        setShowAll(false)
+        setShowNotDone(false)
+        setShowDone(true)
+    }
+
+    function show_not_done() {
+        setShowAll(false)
+        setShowNotDone(true)
+        setShowDone(false)
+    }
+
+    function show_all() {
+        setShowAll(true)
+        setShowNotDone(false)
+        setShowDone(false)
+    }
+
     return (
         <div>
             <h1 className='td-h'>To Do List</h1>
@@ -39,20 +69,46 @@ export default function App() {
                 <input type='text' className='input' placeholder='Add a new item' ref={form}></input>
                 <img className='input-img' src="../images/plus.png" onClick={add_todo}/>
                 <div className='m-div-imgs'>
-                    <img src="../images/done.png"/>
-                    <img src="../images/all.png"/>
-                    <img src="../images/not-done.png"/>
+                    <img src="../images/done.png" onClick={() => show_done()}/>
+                    <img src="../images/all.png" onClick={() => show_all()}/>
+                    <img src="../images/not-done.png" onClick={() => show_not_done()}/>
                 </div>
             </div>
             <h1 className='item-l' ref={item_l}>Item length must be longer than 2.</h1>
             <div>
                 <ul className='m-list'>
                 {todoList.map((name, index) => {
-                    return <div key={index} className='item-div'>
-                                <li>{index + 1}. {name}</li>
-                                <img ref={done} src="images/check.png"/>
-                                <img src="images/delete.jpg" onClick={() => delete_item(index)}/>
-                            </div>
+                    if(showAll) {
+                        if(crossOutList.includes(name)) {
+                            return  <div key={index} className='item-div'>
+                                        <li>{index + 1}. <s>{name}</s></li>
+                                        <img src="images/check.png" onClick={() => cross_item(index)}/>
+                                        <img src="images/delete.jpg" onClick={() => delete_item(index)}/>
+                                    </div>
+                        } else {
+                            return <div key={index} className='item-div'>
+                                        <li>{index + 1}. {name}</li>
+                                        <img src="images/check.png" onClick={() => cross_item(index)}/>
+                                        <img src="images/delete.jpg" onClick={() => delete_item(index)}/>
+                                    </div>
+                            }
+                    } else if(showDone) {
+                        if(crossOutList.includes(name)) {
+                            return  <div key={index} className='item-div'>
+                                        <li>{index + 1}. <s>{name}</s></li>
+                                        <img src="images/check.png" onClick={() => cross_item(index)}/>
+                                        <img src="images/delete.jpg" onClick={() => delete_item(index)}/>
+                                    </div>
+                        }
+                    } else if(showNotDone) {
+                        if(!crossOutList.includes(name)) {
+                            return  <div key={index} className='item-div'>
+                                        <li>{index + 1}. {name}</li>
+                                        <img src="images/check.png" onClick={() => cross_item(index)}/>
+                                        <img src="images/delete.jpg" onClick={() => delete_item(index)}/>
+                                    </div>
+                        }
+                    }
                 })}
                 </ul>
             </div>
